updated gatherTimes.py script
gatherTimes.py creates an xmf metadata file for all the time steps saved in h5. In case of a crash or early termination, the numerical solution xmf will only have the last archived time in it, so this script is used to reconstruct the whole time series. Similar code is called at the end of a simulation so that if a model runs to completion it's not necessary to call this.