The auxiliaryVariables of the domain are now based on a dictionary structure.
This change is based off of work by the HR Wallingford team that is currently on active branches.  This allows auxiliaryVariables to be accessed in more specific ways (allowing us to segregate information to relevant keys on the dictionary).  I hope to use this to intelligently place gauges as part of the auxiliaryVariables.